Update tokens to float

+++ src/palimpzest/dataclasses.py
@@ -16,10 +16,12 @@ class GenerationStats:
     # raw_answers: Optional[List[str]] = field(default_factory=list)
 
     # the total number of input tokens processed by this operator; None if this operation did not use an LLM
-    total_input_tokens: int = 0
+    # typed as a float because GenerationStats may be amortized (i.e. divided) across a number of output records
+    total_input_tokens: float = 0.0
 
     # the total number of output tokens processed by this operator; None if this operation did not use an LLM
-    total_output_tokens: int = 0
+    # typed as a float because GenerationStats may be amortized (i.e. divided) across a number of output records
+    total_output_tokens: float = 0.0
 
     # the total cost of processing the input tokens; None if this operation did not use an LLM
     total_input_cost: float = 0.0
@@ -161,10 +163,12 @@ class RecordOpStats:
     generated_fields: Optional[List[str]] = None
 
     # the total number of input tokens processed by this operator; None if this operation did not use an LLM
-    total_input_tokens: int = 0
+    # typed as a float because GenerationStats may be amortized (i.e. divided) across a number of output records
+    total_input_tokens: float = 0.0
 
     # the total number of output tokens processed by this operator; None if this operation did not use an LLM
-    total_output_tokens: int = 0
+    # typed as a float because GenerationStats may be amortized (i.e. divided) across a number of output records
+    total_output_tokens: float = 0.0
 
     # the total cost of processing the input tokens; None if this operation did not use an LLM
     total_input_cost: float = 0.0
